demo を admin に rename
