+++ src/routes/api/posts/index.js
@@ -1,4 +1,4 @@
-import crud from "$demo/curd.js";
+import crud from "$admin/curd.js";
 
 export const get = crud.index('posts');
 export const post = crud.create('posts');
