+++ src/routes/api/users/index.js
@@ -1,4 +1,4 @@
-import crud from "$demo/curd.js";
+import crud from "$admin/curd.js";
 
 export const get = crud.index('users');
 export const post = crud.create('users');
